--- lightstep/sdk/metric/exporters/otlp/internal/metrictransform/attribute.go
@@ -1,150 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package metrictransform // import "github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/exporters/otlp/internal/metrictransform"
-
-import (
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/sdk/resource"
-	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
-)
-
-// Attributes transforms an attribute.Set into OTLP key-values.
-func Attributes(attrs attribute.Set) []*commonpb.KeyValue {
-	return Iterator(attrs.Iter())
-}
-
-// Iterator transforms an attribute.Set iterator into OTLP key-values.
-func Iterator(iter attribute.Iterator) []*commonpb.KeyValue {
-	l := iter.Len()
-	if l == 0 {
-		return nil
-	}
-
-	out := make([]*commonpb.KeyValue, 0, l)
-	for iter.Next() {
-		out = append(out, KeyValue(iter.Attribute()))
-	}
-	return out
-}
-
-// ResourceAttributes transforms a Resource OTLP key-values.
-func ResourceAttributes(resource *resource.Resource) []*commonpb.KeyValue {
-	return Iterator(resource.Iter())
-}
-
-// KeyValue transforms an attribute KeyValue into an OTLP key-value.
-func KeyValue(kv attribute.KeyValue) *commonpb.KeyValue {
-	return &commonpb.KeyValue{Key: string(kv.Key), Value: Value(kv.Value)}
-}
-
-// Value transforms an attribute Value into an OTLP AnyValue.
-func Value(v attribute.Value) *commonpb.AnyValue {
-	av := new(commonpb.AnyValue)
-	switch v.Type() {
-	case attribute.BOOL:
-		av.Value = &commonpb.AnyValue_BoolValue{
-			BoolValue: v.AsBool(),
-		}
-	case attribute.BOOLSLICE:
-		av.Value = &commonpb.AnyValue_ArrayValue{
-			ArrayValue: &commonpb.ArrayValue{
-				Values: boolSliceValues(v.AsBoolSlice()),
-			},
-		}
-	case attribute.INT64:
-		av.Value = &commonpb.AnyValue_IntValue{
-			IntValue: v.AsInt64(),
-		}
-	case attribute.INT64SLICE:
-		av.Value = &commonpb.AnyValue_ArrayValue{
-			ArrayValue: &commonpb.ArrayValue{
-				Values: int64SliceValues(v.AsInt64Slice()),
-			},
-		}
-	case attribute.FLOAT64:
-		av.Value = &commonpb.AnyValue_DoubleValue{
-			DoubleValue: v.AsFloat64(),
-		}
-	case attribute.FLOAT64SLICE:
-		av.Value = &commonpb.AnyValue_ArrayValue{
-			ArrayValue: &commonpb.ArrayValue{
-				Values: float64SliceValues(v.AsFloat64Slice()),
-			},
-		}
-	case attribute.STRING:
-		av.Value = &commonpb.AnyValue_StringValue{
-			StringValue: v.AsString(),
-		}
-	case attribute.STRINGSLICE:
-		av.Value = &commonpb.AnyValue_ArrayValue{
-			ArrayValue: &commonpb.ArrayValue{
-				Values: stringSliceValues(v.AsStringSlice()),
-			},
-		}
-	default:
-		av.Value = &commonpb.AnyValue_StringValue{
-			StringValue: "INVALID",
-		}
-	}
-	return av
-}
-
-func boolSliceValues(vals []bool) []*commonpb.AnyValue {
-	converted := make([]*commonpb.AnyValue, len(vals))
-	for i, v := range vals {
-		converted[i] = &commonpb.AnyValue{
-			Value: &commonpb.AnyValue_BoolValue{
-				BoolValue: v,
-			},
-		}
-	}
-	return converted
-}
-
-func int64SliceValues(vals []int64) []*commonpb.AnyValue {
-	converted := make([]*commonpb.AnyValue, len(vals))
-	for i, v := range vals {
-		converted[i] = &commonpb.AnyValue{
-			Value: &commonpb.AnyValue_IntValue{
-				IntValue: v,
-			},
-		}
-	}
-	return converted
-}
-
-func float64SliceValues(vals []float64) []*commonpb.AnyValue {
-	converted := make([]*commonpb.AnyValue, len(vals))
-	for i, v := range vals {
-		converted[i] = &commonpb.AnyValue{
-			Value: &commonpb.AnyValue_DoubleValue{
-				DoubleValue: v,
-			},
-		}
-	}
-	return converted
-}
-
-func stringSliceValues(vals []string) []*commonpb.AnyValue {
-	converted := make([]*commonpb.AnyValue, len(vals))
-	for i, v := range vals {
-		converted[i] = &commonpb.AnyValue{
-			Value: &commonpb.AnyValue_StringValue{
-				StringValue: v,
-			},
-		}
-	}
-	return converted
-}
