+++ lightstep/sdk/metric/exporters/otlp/otlpmetric/internal/transform/attribute.go
@@ -12,102 +12,100 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package transform // import "go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal/transform"
+package transform // import "github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/exporters/otlp/otlpmetric/internal/transform"
 
 import (
 	"go.opentelemetry.io/otel/attribute"
-	cpb "go.opentelemetry.io/proto/otlp/common/v1"
+	"go.opentelemetry.io/otel/sdk/resource"
+	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
 )
 
-// AttrIter transforms an attribute iterator into OTLP key-values.
-func AttrIter(iter attribute.Iterator) []*cpb.KeyValue {
+// Attributes transforms an attribute.Set into OTLP key-values.
+func Attributes(attrs attribute.Set) []*commonpb.KeyValue {
+	return Iterator(attrs.Iter())
+}
+
+// Iterator transforms an attribute.Set iterator into OTLP key-values.
+func Iterator(iter attribute.Iterator) []*commonpb.KeyValue {
 	l := iter.Len()
 	if l == 0 {
 		return nil
 	}
 
-	out := make([]*cpb.KeyValue, 0, l)
+	out := make([]*commonpb.KeyValue, 0, l)
 	for iter.Next() {
 		out = append(out, KeyValue(iter.Attribute()))
 	}
 	return out
 }
 
-// KeyValues transforms a slice of attribute KeyValues into OTLP key-values.
-func KeyValues(attrs []attribute.KeyValue) []*cpb.KeyValue {
-	if len(attrs) == 0 {
-		return nil
-	}
-
-	out := make([]*cpb.KeyValue, 0, len(attrs))
-	for _, kv := range attrs {
-		out = append(out, KeyValue(kv))
-	}
-	return out
+// ResourceAttributes transforms a Resource OTLP key-values.
+func ResourceAttributes(resource *resource.Resource) []*commonpb.KeyValue {
+	return Iterator(resource.Iter())
 }
 
 // KeyValue transforms an attribute KeyValue into an OTLP key-value.
-func KeyValue(kv attribute.KeyValue) *cpb.KeyValue {
-	return &cpb.KeyValue{Key: string(kv.Key), Value: Value(kv.Value)}
+func KeyValue(kv attribute.KeyValue) *commonpb.KeyValue {
+	return &commonpb.KeyValue{Key: string(kv.Key), Value: Value(kv.Value)}
 }
 
 // Value transforms an attribute Value into an OTLP AnyValue.
-func Value(v attribute.Value) *cpb.AnyValue {
-	av := new(cpb.AnyValue)
+func Value(v attribute.Value) *commonpb.AnyValue {
+	av := new(commonpb.AnyValue)
 	switch v.Type() {
 	case attribute.BOOL:
-		av.Value = &cpb.AnyValue_BoolValue{
+		av.Value = &commonpb.AnyValue_BoolValue{
 			BoolValue: v.AsBool(),
 		}
 	case attribute.BOOLSLICE:
-		av.Value = &cpb.AnyValue_ArrayValue{
-			ArrayValue: &cpb.ArrayValue{
+		av.Value = &commonpb.AnyValue_ArrayValue{
+			ArrayValue: &commonpb.ArrayValue{
 				Values: boolSliceValues(v.AsBoolSlice()),
 			},
 		}
 	case attribute.INT64:
-		av.Value = &cpb.AnyValue_IntValue{
+		av.Value = &commonpb.AnyValue_IntValue{
 			IntValue: v.AsInt64(),
 		}
 	case attribute.INT64SLICE:
-		av.Value = &cpb.AnyValue_ArrayValue{
-			ArrayValue: &cpb.ArrayValue{
+		av.Value = &commonpb.AnyValue_ArrayValue{
+			ArrayValue: &commonpb.ArrayValue{
 				Values: int64SliceValues(v.AsInt64Slice()),
 			},
 		}
 	case attribute.FLOAT64:
-		av.Value = &cpb.AnyValue_DoubleValue{
+		av.Value = &commonpb.AnyValue_DoubleValue{
 			DoubleValue: v.AsFloat64(),
 		}
 	case attribute.FLOAT64SLICE:
-		av.Value = &cpb.AnyValue_ArrayValue{
-			ArrayValue: &cpb.ArrayValue{
+		av.Value = &commonpb.AnyValue_ArrayValue{
+			ArrayValue: &commonpb.ArrayValue{
 				Values: float64SliceValues(v.AsFloat64Slice()),
 			},
 		}
 	case attribute.STRING:
-		av.Value = &cpb.AnyValue_StringValue{
+		av.Value = &commonpb.AnyValue_StringValue{
 			StringValue: v.AsString(),
 		}
 	case attribute.STRINGSLICE:
-		av.Value = &cpb.AnyValue_ArrayValue{
-			ArrayValue: &cpb.ArrayValue{
+		av.Value = &commonpb.AnyValue_ArrayValue{
+			ArrayValue: &commonpb.ArrayValue{
 				Values: stringSliceValues(v.AsStringSlice()),
 			},
 		}
 	default:
-		av.Value = &cpb.AnyValue_StringValue{
+		av.Value = &commonpb.AnyValue_StringValue{
 			StringValue: "INVALID",
 		}
 	}
 	return av
 }
 
-func boolSliceValues(vals []bool) []*cpb.AnyValue {
-	converted := make([]*cpb.AnyValue, len(vals))
+func boolSliceValues(vals []bool) []*commonpb.AnyValue {
+	converted := make([]*commonpb.AnyValue, len(vals))
 	for i, v := range vals {
-		converted[i] = &cpb.AnyValue{
-			Value: &cpb.AnyValue_BoolValue{
+		converted[i] = &commonpb.AnyValue{
+			Value: &commonpb.AnyValue_BoolValue{
 				BoolValue: v,
 			},
 		}
@@ -115,11 +113,11 @@ func boolSliceValues(vals []bool) []*cpb.AnyValue {
 	return converted
 }
 
-func int64SliceValues(vals []int64) []*cpb.AnyValue {
-	converted := make([]*cpb.AnyValue, len(vals))
+func int64SliceValues(vals []int64) []*commonpb.AnyValue {
+	converted := make([]*commonpb.AnyValue, len(vals))
 	for i, v := range vals {
-		converted[i] = &cpb.AnyValue{
-			Value: &cpb.AnyValue_IntValue{
+		converted[i] = &commonpb.AnyValue{
+			Value: &commonpb.AnyValue_IntValue{
 				IntValue: v,
 			},
 		}
@@ -127,11 +125,11 @@ func int64SliceValues(vals []int64) []*cpb.AnyValue {
 	return converted
 }
 
-func float64SliceValues(vals []float64) []*cpb.AnyValue {
-	converted := make([]*cpb.AnyValue, len(vals))
+func float64SliceValues(vals []float64) []*commonpb.AnyValue {
+	converted := make([]*commonpb.AnyValue, len(vals))
 	for i, v := range vals {
-		converted[i] = &cpb.AnyValue{
-			Value: &cpb.AnyValue_DoubleValue{
+		converted[i] = &commonpb.AnyValue{
+			Value: &commonpb.AnyValue_DoubleValue{
 				DoubleValue: v,
 			},
 		}
@@ -139,11 +137,11 @@ func float64SliceValues(vals []float64) []*cpb.AnyValue {
 	return converted
 }
 
-func stringSliceValues(vals []string) []*cpb.AnyValue {
-	converted := make([]*cpb.AnyValue, len(vals))
+func stringSliceValues(vals []string) []*commonpb.AnyValue {
+	converted := make([]*commonpb.AnyValue, len(vals))
 	for i, v := range vals {
-		converted[i] = &cpb.AnyValue{
-			Value: &cpb.AnyValue_StringValue{
+		converted[i] = &commonpb.AnyValue{
+			Value: &commonpb.AnyValue_StringValue{
 				StringValue: v,
 			},
 		}
