+++ lightstep/sdk/metric/exporters/otlp/otlpmetric/internal/transform/metric_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrictransform
+package transform
 
 import (
 	"math"
